+++ modules/time-sheet/controller/logDetailView.js
